+++ src/main/java/cn/jaylen/codegenerator/service/SysNoticeService.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface SysNoticeService extends BaseService<SysNotice>{
     List<SysNotice> selectAll(SysNoticeExample example);
     Message selectByPages(SysNoticeExample example);
-    int deleteByIDs(Integer[] ids);
+    int deleteByIDs(Long[] ids);
 }
